--- .vscode/launch.json
@@ -1,16 +1,0 @@
-{
-  "version": "0.2.0",
-  "configurations": [
-    {
-      "name": "Python: Flask",
-      "type": "python",
-      "request": "launch",
-      "program": "${workspaceFolder}/app.py",
-      "console": "integratedTerminal",
-      "env": {
-        "FLASK_APP": "app.py",
-        "FLASK_ENV": "development"
-      }
-    }
-  ]
-}
