--- .vscode/settings.json
@@ -1,17 +1,0 @@
-{
-  "sqltools.connections": [
-    {
-      "mysqlOptions": {
-        "authProtocol": "default",
-        "enableSsl": "Disabled"
-      },
-      "previewLimit": 50,
-      "server": "localhost",
-      "port": 3306,
-      "driver": "MySQL",
-      "name": "root",
-      "username": "root",
-      "database": "mydatabase"
-    }
-  ]
-}
