--- AI/requirements.txt
@@ -1,132 +1,0 @@
-anyio==4.4.0
-argon2-cffi==23.1.0
-argon2-cffi-bindings==21.2.0
-arrow==1.3.0
-asttokens @ file:///home/conda/feedstock_root/build_artifacts/asttokens_1698341106958/work
-async-lru==2.0.4
-attrs==24.2.0
-babel==2.16.0
-beautifulsoup4==4.12.3
-bleach==6.1.0
-certifi==2024.7.4
-cffi==1.17.0
-charset-normalizer==3.3.2
-colorama @ file:///home/conda/feedstock_root/build_artifacts/colorama_1666700638685/work
-comm @ file:///home/conda/feedstock_root/build_artifacts/comm_1710320294760/work
-contourpy==1.2.1
-cycler==0.12.1
-debugpy @ file:///D:/bld/debugpy_1722923831604/work
-decorator @ file:///home/conda/feedstock_root/build_artifacts/decorator_1641555617451/work
-defusedxml==0.7.1
-efficientnet_pytorch==0.7.1
-exceptiongroup @ file:///home/conda/feedstock_root/build_artifacts/exceptiongroup_1720869315914/work
-executing @ file:///home/conda/feedstock_root/build_artifacts/executing_1698579936712/work
-fastjsonschema==2.20.0
-filelock==3.15.4
-fonttools==4.53.1
-fqdn==1.5.1
-fsspec==2024.6.1
-h11==0.14.0
-httpcore==1.0.5
-httpx==0.27.0
-idna==3.7
-importlib_metadata @ file:///home/conda/feedstock_root/build_artifacts/importlib-metadata_1721856510709/work
-importlib_resources==6.4.0
-ipykernel @ file:///D:/bld/ipykernel_1719845595208/work
-ipython @ file:///D:/bld/ipython_1701831845989/work
-ipywidgets==8.1.5
-isoduration==20.11.0
-jedi @ file:///home/conda/feedstock_root/build_artifacts/jedi_1696326070614/work
-Jinja2==3.1.4
-json5==0.9.25
-jsonpointer==3.0.0
-jsonschema==4.23.0
-jsonschema-specifications==2023.12.1
-jupyter==1.0.0
-jupyter-console==6.6.3
-jupyter-events==0.10.0
-jupyter-lsp==2.2.5
-jupyter_client @ file:///home/conda/feedstock_root/build_artifacts/jupyter_client_1716472197302/work
-jupyter_core @ file:///D:/bld/jupyter_core_1710257377578/work
-jupyter_server==2.14.2
-jupyter_server_terminals==0.5.3
-jupyterlab==4.2.5
-jupyterlab_pygments==0.3.0
-jupyterlab_server==2.27.3
-jupyterlab_widgets==3.0.13
-kiwisolver==1.4.5
-MarkupSafe==2.1.5
-matplotlib==3.9.1.post1
-matplotlib-inline @ file:///home/conda/feedstock_root/build_artifacts/matplotlib-inline_1713250518406/work
-mistune==3.0.2
-mpmath==1.3.0
-nbclient==0.10.0
-nbconvert==7.16.4
-nbformat==5.10.4
-nest_asyncio @ file:///home/conda/feedstock_root/build_artifacts/nest-asyncio_1705850609492/work
-networkx==3.2.1
-notebook==7.2.1
-notebook_shim==0.2.4
-numpy==1.26.4
-opencv-python==4.10.0.84
-opencv-python-headless==4.10.0.84
-overrides==7.7.0
-packaging @ file:///home/conda/feedstock_root/build_artifacts/packaging_1718189413536/work
-pandas==2.2.2
-pandocfilters==1.5.1
-parso @ file:///home/conda/feedstock_root/build_artifacts/parso_1712320355065/work
-pickleshare @ file:///home/conda/feedstock_root/build_artifacts/pickleshare_1602536217715/work
-pillow==10.4.0
-platformdirs @ file:///home/conda/feedstock_root/build_artifacts/platformdirs_1715777629804/work
-prometheus_client==0.20.0
-prompt_toolkit @ file:///home/conda/feedstock_root/build_artifacts/prompt-toolkit_1718047967974/work
-psutil @ file:///D:/bld/psutil_1719274706746/work
-pure_eval @ file:///home/conda/feedstock_root/build_artifacts/pure_eval_1721585709575/work
-py-cpuinfo==9.0.0
-pycparser==2.22
-Pygments @ file:///home/conda/feedstock_root/build_artifacts/pygments_1714846767233/work
-pyparsing==3.1.2
-python-dateutil @ file:///home/conda/feedstock_root/build_artifacts/python-dateutil_1709299778482/work
-python-json-logger==2.0.7
-pytz==2024.1
-pywin32==306
-pywinpty==2.0.13
-PyYAML==6.0.2
-pyzmq @ file:///D:/bld/pyzmq_1722971783154/work
-qtconsole==5.5.2
-QtPy==2.4.1
-referencing==0.35.1
-requests==2.32.3
-rfc3339-validator==0.1.4
-rfc3986-validator==0.1.1
-rpds-py==0.20.0
-scipy==1.13.1
-seaborn==0.13.2
-Send2Trash==1.8.3
-six @ file:///home/conda/feedstock_root/build_artifacts/six_1620240208055/work
-sniffio==1.3.1
-soupsieve==2.6
-stack-data @ file:///home/conda/feedstock_root/build_artifacts/stack_data_1669632077133/work
-sympy==1.13.1
-terminado==0.18.1
-tinycss2==1.3.0
-tomli==2.0.1
-torch==2.4.0
-torchinfo==1.8.0
-torchvision==0.19.0
-tornado @ file:///D:/bld/tornado_1717722891796/work
-tqdm==4.66.5
-traitlets @ file:///home/conda/feedstock_root/build_artifacts/traitlets_1713535121073/work
-types-python-dateutil==2.9.0.20240821
-typing_extensions @ file:///home/conda/feedstock_root/build_artifacts/typing_extensions_1717802530399/work
-tzdata==2024.1
-ultralytics==8.2.75
-ultralytics-thop==2.0.0
-uri-template==1.3.0
-urllib3==2.2.2
-wcwidth @ file:///home/conda/feedstock_root/build_artifacts/wcwidth_1704731205417/work
-webcolors==24.8.0
-webencodings==0.5.1
-websocket-client==1.8.0
-widgetsnbextension==4.0.13
-zipp @ file:///home/conda/feedstock_root/build_artifacts/zipp_1718013267051/work
