+++ interface-test_v3/Employee.html
@@ -4,7 +4,7 @@
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <title>직원 관리 시스템</title>
-    <link rel="stylesheet" href="/static/css/Employee.css">
+    <link rel="stylesheet" href="Employee.css">
 </head>
 <body>
     <div class="container">
@@ -18,8 +18,9 @@
                 </li>
                 <li><a href="index.html">마이페이지</a></li>
                 <li><a href="index.html">CCTV</a></li>
-                <li><a href="admin.html">관리자페이지</a></li>
+                <li><a href="#">관리자페이지</a></li>
                 <li><a href="#">근골격계 위험 감지</a></li>
+                <li><a href="#">근로자관리</a></li>
                 <li><a href="#">작업장관리</a></li>
                 <li><a href="Employee.html">직원관리</a></li>
             </ul>
@@ -87,6 +88,6 @@
         </div>
     </div>
 
-    <script src="/static/js/Employee.js"></script>
+    <script src="Employee.js"></script>
 </body>
 </html>
