+++ interface-test_v3/login.html
@@ -1,25 +1,27 @@
 <!DOCTYPE html>
 <html lang="ko">
+
 <head>
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <title>로그인</title>
-    <link rel="stylesheet" href="{{ url_for('static', filename='login.css') }}" />
+    <link rel="stylesheet" href="login.css">
 </head>
-<body id="particles-js">
+
+<body>
+
+    <body id="particles-js"></body>
     <div class="animated bounceInDown">
         <div class="container">
+            <span class="error animated tada" id="msg"></span>
             <form name="form1" class="box" onsubmit="return checkStuff()">
                 <h1>로그인</h1>
                 <input type="text" name="아이디" placeholder="아이디">
-                <span class="error" id="idError"></span> <!-- Error message for 아이디 -->
                 <input type="password" name="비밀번호" placeholder="비밀번호">
-                <span class="error" id="passwordError"></span> <!-- Error message for 비밀번호 -->
-                <input type="button" value="로그인" class="btn1" onclick="checkStuff()">
-                <input type="button" value="회원가입" class="btn2" onclick="window.location.href='/signup'">
+                <input type="submit" value="로그인" class="btn1">
+                <input type="button" value="회원가입" class="btn2" onclick="location.href='http://127.0.0.1:5500/sign%20up2/signup2.html'">
             </form>
         </div>
     </div>
-    <script src="{{ url_for('static', filename='login.js') }}"></script>
-</body>
+
 </html>
