+++ interface-test_v3/login.js
@@ -1,0 +1,43 @@
+var pwd = document.getElementById('pwd');
+
+eye.addEventListener('click', togglePass);
+
+function togglePass() {
+
+    eye.classList.toggle('active');
+
+    (pwd.type == 'password') ? pwd.type = 'text' : pwd.type = 'password';
+}
+
+// Form Validation
+
+function checkStuff() {
+    var email = document.form1.email;
+    var password = document.form1.password;
+    var msg = document.getElementById('msg');
+
+    if (email.value == "") {
+        msg.style.display = 'block';
+        msg.innerHTML = "Please enter your email";
+        email.focus();
+        return false;
+    } else {
+        msg.innerHTML = "";
+    }
+
+    if (password.value == "") {
+        msg.innerHTML = "Please enter your password";
+        password.focus();
+        return false;
+    } else {
+        msg.innerHTML = "";
+    }
+    var re = /^(([^<>()[\]\\.,;:\s@\"]+(\.[^<>()[\]\\.,;:\s@\"]+)*)|(\".+\"))@((\[[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}\])|(([a-zA-Z\-0-9]+\.)+[a-zA-Z]{2,}))$/;
+    if (!re.test(email.value)) {
+        msg.innerHTML = "Please enter a valid email";
+        email.focus();
+        return false;
+    } else {
+        msg.innerHTML = "";
+    }
+}
