+++ interface-test_v3/signup.html
@@ -4,8 +4,8 @@
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <title>회원가입 페이지</title>
-    <link href="/static/css/signup.css" rel="stylesheet"/>
-    <script src="/static/js/signup.js" defer></script>
+    <link href="signup.css" rel="stylesheet"/>
+    <script src="signup.js" defer></script>
     <style>
         .modal {
             display: none;
@@ -84,7 +84,7 @@
                     </select>
                 </div>
                 <input type="submit" value="회원가입" class="btn1">
-                <input type="button" value="취소" class="btn2" onclick="location.href='login.html'">
+                <input type="button" value="취소" class="btn2" onclick="location.href='http://127.0.0.1:5500/login/login.html'">
             </form>
         </div>
     </div>
