--- requirements.txt
@@ -1,140 +1,0 @@
-anyio==4.4.0
-argon2-cffi==23.1.0
-argon2-cffi-bindings==21.2.0
-arrow==1.3.0
-asttokens==2.4.1
-async-lru==2.0.4
-attrs==24.2.0
-babel==2.16.0
-beautifulsoup4==4.12.3
-bleach==6.1.0
-blinker==1.8.2
-certifi==2024.7.4
-cffi==1.17.0
-charset-normalizer==3.3.2
-click==8.1.7
-colorama==0.4.6
-comm==0.2.2
-contourpy==1.2.1
-cycler==0.12.1
-debugpy==1.8.5
-decorator==5.1.1
-defusedxml==0.7.1
-efficientnet_pytorch==0.7.1
-exceptiongroup==1.2.2
-executing==2.0.1
-fastjsonschema==2.20.0
-filelock==3.15.4
-Flask==3.0.3
-fonttools==4.53.1
-fqdn==1.5.1
-fsspec==2024.6.1
-h11==0.14.0
-httpcore==1.0.5
-httpx==0.27.0
-idna==3.7
-importlib_metadata==8.2.0
-importlib_resources==6.4.0
-ipykernel==6.29.5
-ipython==8.18.1
-ipywidgets==8.1.5
-isoduration==20.11.0
-itsdangerous==2.2.0
-jedi==0.19.1
-Jinja2==3.1.4
-json5==0.9.25
-jsonpointer==3.0.0
-jsonschema==4.23.0
-jsonschema-specifications==2023.12.1
-jupyter==1.0.0
-jupyter_client==8.6.2
-jupyter-console==6.6.3
-jupyter_core==5.7.2
-jupyter-events==0.10.0
-jupyter-lsp==2.2.5
-jupyter_server==2.14.2
-jupyter_server_terminals==0.5.3
-jupyterlab==4.2.5
-jupyterlab_pygments==0.3.0
-jupyterlab_server==2.27.3
-jupyterlab_widgets==3.0.13
-kiwisolver==1.4.5
-MarkupSafe==2.1.5
-matplotlib==3.9.1.post1
-matplotlib-inline==0.1.7
-mistune==3.0.2
-mpmath==1.3.0
-mysql-connector-python==9.0.0
-nbclient==0.10.0
-nbconvert==7.16.4
-nbformat==5.10.4
-nest_asyncio==1.6.0
-networkx==3.2.1
-notebook==7.2.1
-notebook_shim==0.2.4
-numpy==1.26.4
-opencv-python==4.10.0.84
-opencv-python-headless==4.10.0.84
-overrides==7.7.0
-packaging==24.1
-pandas==2.2.2
-pandocfilters==1.5.1
-parso==0.8.4
-pickleshare==0.7.5
-pillow==10.4.0
-platformdirs==4.2.2
-prometheus_client==0.20.0
-prompt_toolkit==3.0.47
-psutil==6.0.0
-pure_eval==0.2.3
-py-cpuinfo==9.0.0
-pycparser==2.22
-Pygments==2.18.0
-pyparsing==3.1.2
-python-dateutil==2.9.0
-python-json-logger==2.0.7
-pytz==2024.1
-pywin32==306
-pywinpty==2.0.13
-PyYAML==6.0.2
-pyzmq==26.1.0
-qtconsole==5.5.2
-QtPy==2.4.1
-referencing==0.35.1
-requests==2.32.3
-rfc3339-validator==0.1.4
-rfc3986-validator==0.1.1
-rpds-py==0.20.0
-scipy==1.13.1
-seaborn==0.13.2
-Send2Trash==1.8.3
-setuptools==72.1.0
-six==1.16.0
-sniffio==1.3.1
-soupsieve==2.6
-stack-data==0.6.2
-sympy==1.13.1
-terminado==0.18.1
-tinycss2==1.3.0
-tomli==2.0.1
-torch==2.4.0
-torchinfo==1.8.0
-torchvision==0.19.0
-tornado==6.4.1
-tqdm==4.66.5
-traitlets==5.14.3
-types-python-dateutil==2.9.0.20240821
-typing_extensions==4.12.2
-tzdata==2024.1
-ultralytics==8.2.75
-ultralytics-thop==2.0.0
-uri-template==1.3.0
-urllib3==2.2.2
-wcwidth==0.2.13
-webcolors==24.8.0
-webencodings==0.5.1
-websocket-client==1.8.0
-Werkzeug==3.0.4
-wheel==0.43.0
-widgetsnbextension==4.0.13
-zipp==3.19.2
