--- static/css/admin.css
@@ -1,205 +1,0 @@
-body {
-    font-family: Arial, sans-serif;
-    margin: 0;
-    padding: 0;
-    background-color: #f0f0f0;
-}
-
-.container {
-    display: flex;
-    height: 100vh;
-    background-color: #212121;
-}
-
-.sidebar {
-    width: 200px;
-    background-color: #007BFF;
-    padding-top: 20px;
-    border-right: 1px solid #ccc;
-    display: flex;
-    flex-direction: column;
-    justify-content: space-between;
-}
-
-.sidebar ul {
-    list-style-type: none;
-    padding: 0;
-    margin: 0;
-}
-
-.sidebar ul li {
-    margin: 10px 0;
-    border-bottom: 1px solid #0056b3;
-}
-
-.sidebar ul li a {
-    color: white;
-    text-decoration: none;
-    display: block;
-    padding: 10px 20px;
-    border-radius: 4px;
-}
-
-.sidebar ul li a:hover {
-    background-color: #0056b3;
-}
-
-#logoutButton {
-    margin: 20px;
-    padding: 10px;
-    background-color: #0056b3;
-    color: white;
-    border: none;
-    border-radius: 4px;
-    cursor: pointer;
-}
-
-#logoutButton:hover {
-    background-color: #004494;
-}
-
-.main-content {
-    flex: 1;
-    padding: 20px;
-    overflow-y: auto;
-}
-
-.main-header {
-    background-color: #007BFF;
-    color: white;
-    padding: 10px 20px;
-    border-radius: 4px;
-    margin-bottom: 20px;
-    text-align: center;
-}
-
-.system-title {
-    margin: 0;
-}
-
-.content {
-    display: flex;
-    flex-direction: column;
-    align-items: flex-start; /* 왼쪽 정렬로 변경 */
-}
-
-.action-bar {
-    margin-bottom: 10px;
-}
-
-#employeeTable {
-    width: 100%;
-    border-collapse: collapse;
-    margin-bottom: 20px;
-    background-color: white; /* 흰색 배경으로 변경 */
-    min-height: 600px; /* 최소 높이 추가 */
-    padding-bottom: 20px; /* 아래 패딩 추가 */
-}
-
-#employeeTable, #employeeTable th, #employeeTable td {
-    border: 1px solid #ccc;
-}
-
-#employeeTable th, #employeeTable td {
-    padding: 10px;
-    text-align: left;
-}
-
-.btn {
-    background-color: #4CAF50;
-    color: white;
-    padding: 10px 20px;
-    border: none;
-    cursor: pointer;
-    border-radius: 4px;
-}
-
-.btn:hover {
-    background-color: #45a049;
-}
-
-.btn.small-btn {
-    padding: 5px 10px; /* 크기를 줄임 */
-}
-
-.modal {
-    display: none;
-    position: fixed;
-    z-index: 1;
-    left: 0;
-    top: 0;
-    width: 100%;
-    height: 100%;
-    overflow: auto;
-    background-color: rgb(0,0,0);
-    background-color: rgba(0,0,0,0.4);
-    padding-top: 60px;
-}
-
-.modal-content {
-    background-color: #fefefe;
-    margin: 5% auto;
-    padding: 20px;
-    border: 1px solid #888;
-    width: 80%;
-}
-
-.close {
-    color: #aaa;
-    float: right;
-    font-size: 28px;
-    font-weight: bold;
-}
-
-.close:hover,
-.close:focus {
-    color: black;
-    text-decoration: none;
-    cursor: pointer;
-}
-
-input[type="text"], input[type="email"], select {
-    width: 100%;
-    padding: 12px 20px;
-    margin: 8px 0;
-    display: inline-block;
-    border: 1px solid #ccc;
-    border-radius: 4px;
-    box-sizing: border-box;
-}
-
-select {
-    height: auto; /* 높이 자동으로 조정 */
-    padding: 8px 12px; /* 패딩 조정 */
-    text-align: center; /* 텍스트 중앙 정렬 */
-}
-
-.user-info {
-    display: flex;
-    justify-content: center; /* 수평으로 중앙 정렬 */
-    align-items: center; /* 수직으로 중앙 정렬 */
-    color: white;
-}
-
-.notification-button {
-    background-color: #0056b3; /* 알림 버튼 색상 */
-    color: white;
-    border: none;
-    border-radius: 5px;
-    padding: 10px;
-    margin: 10px;
-    cursor: pointer;
-    position: relative; /* 알림 카운트 위치를 조정하기 위해 필요 */
-}
-
-.notification-count {
-    background-color: red; /* 알림 카운트 배경 색상 */
-    color: white;
-    border-radius: 50%;
-    padding: 5px 10px;
-    position: absolute;
-    top: -10px;
-    right: -10px;
-    font-size: 12px;
-    font-weight: bold;
-}
