--- static/login.css
@@ -1,191 +1,0 @@
-body,
-html {
-    font-family: 'Source Sans Pro', sans-serif;
-    background-color: #f1ecec;
-    padding: 0;
-    margin: 0;
-}
-
-#particles-js {
-    position: absolute;
-    width: 100%;
-    height: 100%;
-}
-
-.container {
-    margin: 0;
-    top: 50px;
-    left: 50%;
-    position: absolute;
-    text-align: center;
-    transform: translateX(-50%);
-    background-color: #fff;
-    border-radius: 9px;
-    border-top: 10px solid #CBDDF5;
-    border-bottom: 10px solid #CBDDF5;
-    width: 400px;
-    height: 500px;
-}
-
-.box h1 {
-    font-family: 'Source Sans Pro', sans-serif;
-    color: #020202;
-    font-size: 18px;
-    margin-top: 94px;
-    ;
-}
-
-.box h4 span {
-    color: #dfdeee;
-    font-weight: lighter;
-}
-
-.box h5 {
-    font-family: 'Source Sans Pro', sans-serif;
-    font-size: 13px;
-    color: #a1a4ad;
-    letter-spacing: 1.5px;
-    margin-top: -15px;
-    margin-bottom: 70px;
-}
-
-.box input[type="text"],
-.box input[type="password"] {
-    display: block;
-    margin: 20px auto;
-    background: #fff;
-    border: 1px solid #ccc;
-    border-radius: 5px;
-    padding: 14px 10px;
-    width: 320px;
-    outline: none;
-    color: #000000;
-    -webkit-transition: all .2s ease-out;
-    -moz-transition: all .2s ease-out;
-    -ms-transition: all .2s ease-out;
-    -o-transition: all .2s ease-out;
-    transition: all .2s ease-out;
-
-}
-
-::-webkit-input-placeholder {
-    color: #565f79;
-}
-
-.box input[type="text"]:focus,
-.box input[type="password"]:focus {
-    border: 1px solid #79A6FE;
-
-}
-
-a {
-    color: #5c7fda;
-    text-decoration: none;
-}
-
-a:hover {
-    text-decoration: underline;
-}
-
-label input[type="checkbox"] {
-    display: none;
-    /* hide the default checkbox */
-}
-
-/* style the artificial checkbox */
-label span {
-    height: 13px;
-    width: 13px;
-    border: 2px solid #fff;
-    border-radius: 2px;
-    display: inline-block;
-    position: relative;
-    cursor: pointer;
-    float: left;
-    left: 7.5%;
-}
-
-.forgetpass {
-    display: block;
-    margin-bottom: 60px;
-    text-align: right;
-    color: #666;
-    text-decoration: none;
-}
-
-.btn1,
-.btn2 {
-    background-color: #4B89DC;
-    color: #fff;
-    border: none;
-    padding: 10px 20px;
-    text-align: center;
-    text-decoration: none;
-    display: block;
-    font-size: 16px;
-    margin: 10px auto;
-    cursor: pointer;
-    border-radius: 25px;
-    width: 80%;
-}
-
-.rmb {
-    position: absolute;
-    margin-left: -24%;
-    margin-top: 0px;
-    color: #dfdeee;
-    font-size: 13px;
-}
-
-.forgetpass {
-    position: relative;
-    float: right;
-    right: 28px;
-}
-
-.dnthave {
-    position: absolute;
-    top: 92%;
-    left: 24%;
-}
-
-[type=checkbox]:checked+span:before {
-    /* <-- style its checked state */
-    font-family: FontAwesome;
-    font-size: 16px;
-    content: "\f00c";
-    position: absolute;
-    top: -4px;
-    color: #896cec;
-    left: -1px;
-    width: 13px;
-}
-
-.typcn {
-    position: absolute;
-    left: 339px;
-    top: 282px;
-    color: #3b476b;
-    font-size: 22px;
-    cursor: pointer;
-}
-
-.typcn.active {
-    color: #7f60eb;
-}
-
-.error {
-    background: #ff3333;
-    text-align: center;
-    width: 337px;
-    height: 20px;
-    padding: 2px;
-    border: 0;
-    border-radius: 5px;
-    margin: 10px auto 10px;
-    position: absolute;
-    top: 31%;
-    left: 7.2%;
-    color: white;
-    display: none;
-}
