--- static/signup.css
@@ -1,242 +1,0 @@
-body, html {
-    font-family: 'Source Sans Pro', sans-serif;
-    background-color: #f1ecec;
-    padding: 0;
-    margin: 0;
-}
-
-/* 파티클 백그라운드 */
-#particles-js {
-    position: absolute;
-    width: 100%;
-    height: 100%;
-}
-
-.container {
-    margin: 0;
-    top: 35px;
-    left: 50%;
-    position: absolute;
-    text-align: center;
-    transform: translateX(-50%);
-    background-color: #fff;
-    border-radius: 9px;
-    border-top: 10px solid #CBDDF5;
-    border-bottom: 10px solid #CBDDF5;
-    width: 500px;
-    height: 800px;
-}
-
-/* 제목 스타일 */
-.box h1 {
-    color: #ffffff; 
-    font-size: 25px;
-    margin-top: 70px;
-}
-
-.box h1 span {
-    color: #000000;
-    font-weight: lighter;
-}
-
-/* 서브 제목 스타일 */
-.box h2 {
-    font-size: 13px;
-    color: #000000;
-    letter-spacing: 1.5px;
-    margin: 20px auto;
-}
-
-input[type="text"], input[type="password"], input[type="tel"] {
-    display: block;
-    margin: 5px auto;
-    margin-bottom: 20px;
-    background: #ffffff;
-    border: 1px solid #ccc;
-    border-radius: 5px;
-    padding: 14px 10px;
-    width: 320px;
-    outline: none;
-    color: #000000;
-    transition: all .2s ease-out;
-}
-
-input[type="text"]:focus, input[type="password"]:focus {
-    border: 1px solid #79A6FE;
-}
-
-.id-container {
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    gap: 10px;
-    margin: 10px auto;
-    width: fit-content;
-}
-
-
-.id-container input[type="text"] {
-    width: 220px;
-    text-align: left;
-    margin: 0;
-    box-sizing: border-box;
-}
-
-.id-container .btn3 {
-    margin: 0;
-}
-
-
-.btn1, .btn2 {
-    margin-top: 20px;
-    padding: 10px 20px;
-    background-color: #4B89DC;
-    color: #fff;
-    border: none;
-    border-radius: 5px;
-    cursor: pointer;
-    font-size: 16px;
-}
-
-.btn3 {
-    padding: 10px 20px;
-    background-color: #4B89DC;
-    color: #fff;
-    border: none;
-    border-radius: 5px;
-    cursor: pointer;
-    font-size: 15px;
-    white-space: nowrap;
-    min-width: 100px;
-}
-
-
-.btn1:hover, .btn2:hover, .btn3:hover {
-    background-color: #0056b3;
-}
-
-.position select, .email-container select {
-    display: block;
-    margin: 20px auto;
-    background: #ffffff;
-    border: 1px solid #ccc;
-    border-radius: 5px;
-    padding: 14px 10px;
-    width: 320px;
-    color: #000000;
-    outline: none;
-    transition: all .2s ease-out;
-}
-
-.position select:focus, .email-container select:focus {
-    border: 1px solid #79A6FE;
-}
-
-a {
-    color: #5c7fda;
-    text-decoration: none;
-}
-
-a:hover {
-    text-decoration: underline;
-}
-
-/* 에러 메시지 스타일 */
-.error-text {
-    color: red;
-    font-size: 14px;
-    margin-top: 3px;
-    display: block;
-}
-
-.footer {
-    position: relative;
-    left: 0;
-    bottom: 0;
-    top: 605px;
-    width: 100%;
-    color: #78797d;
-    font-size: 14px;
-    text-align: center;
-}
-
-.footer .fa {
-    color: #7f5feb;
-}
-
-.modal {
-    display: none;
-    position: fixed;
-    z-index: 1;
-    left: 0;
-    top: 0;
-    width: 100%;
-    height: 100%;
-    overflow: auto;
-    background-color: rgba(0, 0, 0, 0.4);
-    padding-top: 60px;
-}
-
-.modal-content {
-    background-color: #fefefe;
-    margin: 5% auto;
-    padding: 20px;
-    border: 1px solid #888;
-    width: 80%;
-    max-width: 300px;
-    text-align: center;
-}
-
-.close {
-    color: #aaa;
-    float: right;
-    font-size: 28px;
-    font-weight: bold;
-}
-
-.close:hover, .close:focus {
-    color: black;
-    text-decoration: none;
-    cursor: pointer;
-}
-
-.phone-container, .email-container {
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    margin: 20px 0;
-}
-
-.phone-container input {
-    padding-left: 10px;
-    border: 1px solid #ccc;
-    border-radius: 5px;
-    text-align: center;
-    box-sizing: border-box;
-}
-
-.phone-container input[type="text"] {
-    width: 60px;
-    padding: 14px 10px;
-    margin: 0 15px;
-}
-
-.email-container input[type="text"] {
-    width: 150px; 
-    padding: 14px 10px;
-    margin: 0 10px;
-    border: 1px solid #ccc;
-    border-radius: 5px;
-    text-align: center; 
-    box-sizing: border-box;
-}
-
-.phone h2 {
-    font-size: 14px;
-    color: #000;
-    margin: 0;
-    padding: 0;
-    position: absolute; 
-    left: 22%;
-    transform: translateX(-50%); 
-}
