--- templates/Employee.html
@@ -1,116 +1,0 @@
-<!DOCTYPE html>
-<html lang="ko">
-  <head>
-    <meta charset="UTF-8" />
-    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <title>직원 관리 시스템</title>
-    <link
-      rel="stylesheet"
-      href="{{ url_for('static', filename='Employee.css') }}"
-    />
-  </head>
-  <body>
-    <div class="container">
-      <nav class="sidebar">
-        <ul>
-          <li class="user-info">
-            <span id="userName" class="user-name">{{ member_name }}님</span>
-            <button id="notificationButton" class="notification-button">
-              알림<span id="notificationCount" class="notification-count"
-                >3</span
-              >
-            </button>
-          </li>
-          <li><a href="/">마이페이지</a></li>
-          <li><a href="/index">CCTV</a></li>
-          <li><a href="/index">관리자페이지</a></li>
-          <li><a href="#">근골격계 위험 감지</a></li>
-          <li><a href="#">근로자관리</a></li>
-          <li><a href="#">작업장관리</a></li>
-          <li><a href="/Employee">직원관리</a></li>
-        </ul>
-        <button id="logoutButton">로그아웃</button>
-      </nav>
-      <main class="main-content">
-        <header class="main-header">
-          <h1 class="system-title">직원 관리 시스템</h1>
-        </header>
-        <section class="content">
-          <div class="action-bar">
-            <button id="addEmployeeButton" class="btn small-btn">
-              직원 추가
-            </button>
-          </div>
-          <table id="employeeTable" class="white-background">
-            <thead>
-              <tr>
-                <th>이름</th>
-                <th>지역</th>
-                <th>부서</th>
-                <th>직급</th>
-                <th>전화번호</th>
-                <th>이메일</th>
-                <th>수정</th>
-                <th>삭제</th>
-              </tr>
-            </thead>
-            <tbody>
-              {% for employee in employees %}
-              <tr>
-                <td>{{ employee[0] }}</td>
-                <!-- 이름 -->
-                <td>{{ employee[1] }}</td>
-                <!-- 지역 -->
-                <td>{{ employee[2] }}</td>
-                <!-- 부서 -->
-                <td>{{ employee[3] }}</td>
-                <!-- 직급 -->
-                <td>{{ employee[4] }}</td>
-                <!-- 전화번호 -->
-                <td>{{ employee[5] }}</td>
-                <!-- 이메일 -->
-                <td><button class="btn small-btn">수정</button></td>
-                <td><button class="btn small-btn">삭제</button></td>
-              </tr>
-              {% endfor %}
-            </tbody>
-          </table>
-        </section>
-      </main>
-    </div>
-
-    <div id="employeeModal" class="modal">
-      <div class="modal-content">
-        <span class="close">&times;</span>
-        <h2 id="modalTitle">직원 추가하기</h2>
-        <form id="employeeForm" method="POST" action="/Employee">
-          <label for="name">이름:</label>
-          <input type="text" id="name" name="name" required /><br />
-          <label for="region">지역:</label>
-          <input type="text" id="region" name="region" required /><br />
-          <label for="department">부서:</label>
-          <select id="department" name="department" required>
-            <option value="부서1">부서1</option>
-            <option value="부서2">부서2</option>
-            <option value="부서3">부서3</option>
-            <option value="부서4">부서4</option></select
-          ><br />
-          <label for="position">직급:</label>
-          <select id="position" name="position" required>
-            <option value="1">1</option>
-            <option value="2">2</option>
-            <option value="3">3</option>
-            <option value="4">4</option></select
-          ><br />
-          <label for="phone">전화번호:</label>
-          <input type="text" id="phone" name="phone" required /><br />
-          <label for="email">이메일:</label>
-          <input type="email" id="email" name="email" required /><br />
-          <button type="submit" class="btn small-btn">저장</button>
-        </form>
-      </div>
-    </div>
-
-    <script src="{{ url_for('static', filename='Employee.js') }}"></script>
-  </body>
-</html>
