--- templates/login.html
@@ -1,30 +1,0 @@
-<!DOCTYPE html>
-<html lang="ko">
-
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>로그인</title>
-    <link
-      rel="stylesheet"
-      href="{{ url_for('static', filename='login.css') }}"
-    />
-</head>
-
-<body>
-
-    <body id="particles-js"></body>
-    <div class="animated bounceInDown">
-        <div class="container">
-            <span class="error animated tada" id="msg"></span>
-            <form name="form1" class="box" onsubmit="return checkStuff()"method="POST" action="http://localhost:5000/">
-                <h1>로그인</h1>
-                <input type="text" name="아이디" placeholder="아이디">
-                <input type="password" name="비밀번호" placeholder="비밀번호">
-                <input type="submit" value="로그인" class="btn1" onclick="location.href='/index'">
-                <input type="button" value="회원가입" class="btn2" onclick="location.href='/signup'">
-            </form>
-        </div>
-    </div>
-
-</html>
