--- templates/register.html
@@ -1,85 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-  <head>
-    <meta charset="utf-8" />
-    <meta
-      name="viewport"
-      content="width=device-width, initial-scale=1, shrink-to-fit=no"
-    />
-    <link
-      rel="stylesheet"
-      href="https://stackpath.bootstrapcdn.com/bootstrap/4.4.1/css/bootstrap.min.css"
-      integrity="sha384-Vkoo8x4CGsO3+Hhxv8T/Q5PaXtkKtu6ug5TOeNV6gBiFeWPGFN9MuhOf23Q9Ifjh"
-      crossorigin="anonymous"
-    />
-    <script
-      src="https://code.jquery.com/jquery-3.4.1.slim.min.js"
-      integrity="sha384-J6qa4849blE2+poT4WnyKhv5vZF5SrPo0iEjwBvKU7imGFAV0wwj1yYfoRSJoZ+n"
-      crossorigin="anonymous"
-    ></script>
-    <script
-      src="https://cdn.jsdelivr.net/npm/popper.js@1.16.0/dist/umd/popper.min.js"
-      integrity="sha384-Q6E9RHvbIyZFJoft+2mJbHaEWldlvI9IOYy5n3zV9zzTtmI3UksdQRVvoxMfooAo"
-      crossorigin="anonymous"
-    ></script>
-    <script
-      src="https://stackpath.bootstrapcdn.com/bootstrap/4.4.1/js/bootstrap.min.js"
-      integrity="sha384-wfSDF2E50Y2D1uUdj0O3uMBJnjuUD4Ih7YwaYd1iqfktj0Uod8GCExl3Og8ifwB6"
-      crossorigin="anonymous"
-    ></script>
-  </head>
-  <body>
-    <div class="container">
-      <div class="row mt-5">
-        <h1>회원가입</h1>
-      </div>
-
-      <div class="col-12">
-        <br /><br />
-        <form method="POST" action=".">
-          <div class="form-group">
-            <label for="userid">아이디</label>
-            <input
-              type="text"
-              class="form-control"
-              id="userid"
-              placeholder="아이디"
-              name="userid"
-            />
-          </div>
-          <div class="form-group">
-            <label for="email">이메일</label>
-            <input
-              type="email"
-              class="form-control"
-              id="email"
-              placeholder="이메일"
-              name="email"
-            />
-          </div>
-          <div class="form-group">
-            <label for="password">비밀번호</label>
-            <input
-              type="password"
-              class="form-control"
-              id="password"
-              placeholder="비밀번호"
-              name="password"
-            />
-          </div>
-          <div class="form-group">
-            <label for="re_password">비밀번호 확인</label>
-            <input
-              type="password"
-              class="form-control"
-              id="re_password"
-              placeholder="비밀번호 확인"
-              name="re_password"
-            />
-          </div>
-          <button type="submit" class="btn btn-primary">제출</button>
-        </form>
-      </div>
-    </div>
-  </body>
-</html>
