--- templates/signup.html
@@ -1,145 +1,0 @@
-<!DOCTYPE html>
-<html lang="ko">
-  <head>
-    <meta charset="UTF-8" />
-    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <title>회원가입 페이지</title>
-    <link
-      rel="stylesheet"
-      href="{{ url_for('static', filename='signup.css') }}"
-    />
-    <script src="/static/js/signup.js" defer></script>
-    <style>
-      .modal {
-        display: none;
-        position: fixed;
-        z-index: 1;
-        left: 0;
-        top: 0;
-        width: 100%;
-        height: 100%;
-        overflow: auto;
-        background-color: rgba(0, 0, 0, 0.4);
-        padding-top: 60px;
-      }
-      .modal-content {
-        background-color: #fefefe;
-        margin: 5% auto;
-        padding: 20px;
-        border: 1px solid #888;
-        width: 80%;
-        max-width: 300px;
-        text-align: center;
-      }
-      .close {
-        color: #aaa;
-        float: right;
-        font-size: 28px;
-        font-weight: bold;
-      }
-      .close:hover,
-      .close:focus {
-        color: black;
-        text-decoration: none;
-        cursor: pointer;
-      }
-      .error-text {
-        color: red;
-      }
-
-      .btn-id-check {
-        padding: 10px 20px;
-        background-color: #4B89DC;
-        color: #fff;
-        border: none;
-        border-radius: 5px;
-        cursor: pointer;
-        font-size: 14px;
-      }
-      .btn-id-check:hover {
-        background-color: #0056b3;
-      }
-    </style>
-  </head>
-  <body id="particles-js">
-    <div class="animated bounceInDown">
-      <div class="container">
-        <form name="form1" class="box" method="POST" action="http://localhost:5000/signup">
-          <h1><span>회원가입</span></h1>
-          <div class="id-container">
-            <input
-              type="text"
-              name="id"
-              placeholder="아이디 입력(6~20자)"
-              autocomplete="on"
-            />
-            <button type="button" class="btn3" onclick="checkUsername()">중복 확인</button>
-            <div id="usernameCheckResult" class="error-text"></div>
-          </div>
-          <input
-            type="password"
-            name="pw"
-            placeholder="비밀번호 입력(문자, 숫자, 특수문자 포함(8~20자))"
-            id="pwd1"
-            autocomplete="off"
-          />        
-          <input
-            type="password"
-            name="pw_ch"
-            placeholder="비밀번호 재입력"
-            id="pwd2"
-            autocomplete="off"
-          />
-          <input
-            type="text"
-            name="name"
-            placeholder="이름을 입력해 주세요"
-            autocomplete="on"
-          />
-          <input
-            type="text"
-            name="tel"
-            placeholder="전화번호를 입력해 주세요"
-            autocomplete="on"
-          />
-          <input
-            type="text"
-            name="email"
-            placeholder="이메일을 입력해 주세요"
-            autocomplete="on"
-          />
-          <input
-            type="text"
-            name="address"
-            placeholder="주소를 입력해 주세요(시, 구)"
-            autocomplete="on"
-          />
-          <span id="msg" class="error-text" style="display: none"></span>
-          <div class="position">
-            <select id="position" name="position">
-              <option selected disabled>직책</option>
-              <option>1</option>
-              <option>2</option>
-              <option>3</option>
-              <option>4</option>
-            </select>
-          </div>
-          <input type="submit" value="회원가입" class="btn1" />
-          <input
-            type="submit"
-            value="취소"
-            class="btn2"
-            onclick="location.href='http://127.0.0.1:5000/'"
-          />
-        </form>
-      </div>
-    </div>
-
-    <div id="myModal" class="modal">
-      <div class="modal-content">
-        <span class="close">&times;</span>
-        <p>가입이 완료되었습니다.</p>
-      </div>
-    </div>
-  </body>
-</html>
